requirements.txt
PyMuPDF>=1.24.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.33.0
pandas>=2.0.0
